move assert_raises to util
